Use fixed width integer types in Compiler (#854)
* Use fixed width integer types in Compiler

- Contributes to #535
- Contributes to #531
- Also address many code analysis issues from VS19 in-IDE parser
  - e.g. Use of enum class in preference to unscoped enums
- Define specific integer types (using enum class) for IPs and text positions to help identify
errors

* Post self code-review cleanup

* Wrong max radix

+++ Core/DolphinVM/InProcStub/StdAfx.h
@@ -13,8 +13,10 @@
 #define _ATL_DEBUG_INTERFACES
 #endif
 
+#include <stdint.h>
+
 #define UMDF_USING_NTSTATUS
-typedef long NTSTATUS;
+typedef int32_t NTSTATUS;
 #include <ntstatus.h>
 
 #include <atlbase.h>
